fixed small error, improved interface
