+++ lib/userInfo/ExtraInfo.dart
@@ -1,0 +1,14 @@
+
+
+class ExtraInfo{
+
+  String infoType = '';
+  String info = '';
+
+  ExtraInfo(this.infoType,  this.info);
+
+  ExtraInfo.fromExtraInfo(ExtraInfo otherExtraInfo){
+    this.infoType = otherExtraInfo.infoType;
+    this.info = otherExtraInfo.info ;
+  }
+}
