+++ test/widget_test.dart
@@ -9,6 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:projeto_esof/main.dart';
+import 'package:projeto_esof/userInfo/screens.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
